+++ src/rics/translation/factory.py
@@ -1,6 +1,5 @@
 """Factory functions for translation classes."""
 
-import warnings
 from typing import TYPE_CHECKING, Any, Callable, Dict, Iterable, Optional, Tuple, Union
 
 import toml
@@ -14,30 +13,53 @@ from rics.utility.collections import InheritedKeysDict
 from rics.utility.collections.inherited_keys_dict import DefaultType, MakeDict, SpecificType
 
 if TYPE_CHECKING:
-    from rics.translation._translator import Translator  # pragma: no cover
+    from rics.translation._translator import Translator
 
-FetcherFactory = Callable[[str, Dict[str, Any]], Fetcher]
+FetcherFactory = Callable[[str, Dict[str, Any]], AbstractFetcher]
 MakeFetcherType = Union[Fetcher, FetcherFactory]
 MapperFactory = Callable[[Dict[str, Any], bool], Optional[Mapper]]
 MakeMapperType = Union[Mapper, MapperFactory]
 
 
-def default_fetcher_factory(name: str, kwargs: Dict[str, Any]) -> Fetcher:
-    """Create a Fetcher from a dict config."""
-    fetcher_clazz = getattr(fetching, name, None)
+def default_fetcher_factory(clazz: str, config: Dict[str, Any]) -> AbstractFetcher:
+    """Create a Fetcher from a dict config.
+
+    Args:
+        clazz: Type of Fetcher to create.
+        config: Keyword arguments for the fetcher.
+
+    Returns:
+        An AbstractFetcher instance.
+
+    Raises:
+        ConfigurationError: If `config` is invalid.
+        TypeError: If `clazz` is not an AbstractFetcher subtype.
+    """
+    fetcher_clazz = getattr(fetching, clazz, None)
     if fetcher_clazz is None:
-        raise ValueError(f"Fetcher class '{name}' not known. Consider using the 'fetcher_factory' argument.")
-    fetcher = fetcher_clazz(**kwargs)
-    if not isinstance(fetcher, Fetcher):
+        raise ConfigurationError(
+            f"Fetcher class '{clazz}' not known. Use the 'fetcher_factory' argument "
+            "to initialize customer implementations."
+        )
+    fetcher = fetcher_clazz(**config)
+    if not isinstance(fetcher, AbstractFetcher):
         raise TypeError(fetcher)
     return fetcher
 
 
 def default_mapper_factory(config: Dict[str, Any], for_fetcher: bool) -> Optional[Mapper]:
-    """Create a Mapper from a dict config."""
-    if not config:
-        return None
+    """Create a Mapper from a dict config.
 
+    Args:
+        config: Keyword arguments for the fetcher.
+        for_fetcher: Flag indicating that the mapper returned will be used by an AbstractFetcher instance.
+
+    Returns:
+        A Mapper instance.
+
+    Raises:
+        ConfigurationError: If `config` is invalid.
+    """
     if "score_function" in config and not isinstance(config["score_function"], str):
         score_function = config.pop("score_function")
 
@@ -85,8 +107,8 @@ def translator_from_toml_config(
         extra_fetchers: Path to TOML files defining additional fetchers. Useful for fetching from multiple sources or
             kinds of sources, for example locally stored files in conjunction with one or more databases. The fetchers
             are ranked by input order, with the fetcher defined in `file` being given the highest priority (rank 0).
-        fetcher_factory: A pre-initialized Fetcher, or a callable taking (name, kwargs) which returns a Fetcher.
-        mapper_factory: A pre-initialized Mapper, or a callable taking (kwargs) which returns a Mapper. Used for both
+        fetcher_factory: A Fetcher instance, or a callable taking (name, kwargs) which returns an AbstractFetcher.
+        mapper_factory: A Mapper instance, or a callable taking (kwargs) which returns a Mapper. Used for both
             Translator and Fetcher mapper initialization.
 
     Returns:
@@ -105,17 +127,8 @@ def translator_from_toml_config(
     _check_allowed_keys(["translator", "mapping", "fetching", "unknown_ids"], config, "<root>")
 
     translator_config = config.pop("translator", {})
-    fetcher = _make_fetcher(fetcher_factory, mapper_factory, **config.pop("fetching", {}))
-    if extra_fetchers:
-        fetcher = MultiFetcher(
-            fetcher,
-            *(
-                _make_fetcher(fetcher_factory, mapper_factory, **toml.load(file_fetcher_file)["fetching"])
-                for file_fetcher_file in extra_fetchers
-            ),
-            # Note: Source discovery isn't done until mapping is done in the Translator
-            duplicate_source_discovered_action="warn",
-        )
+
+    fetcher = _handle_fetching(config.pop("fetching", {}), extra_fetchers, fetcher_factory, mapper_factory)
 
     mapper = _make_mapper("translator", mapper_factory, translator_config)
     default_fmt, default_translations = _make_default_translations(**config.pop("unknown_ids", {}))
@@ -123,6 +136,33 @@ def translator_from_toml_config(
     return Translator(
         fetcher, mapper=mapper, default_translations=default_translations, default_fmt=default_fmt, **translator_config
     )
+
+
+def _handle_fetching(
+    config: Dict[str, Any],
+    extra_fetchers: Iterable[str],
+    fetcher_factory: MakeFetcherType,
+    mapper_factory: MakeMapperType,
+) -> Fetcher:
+    if isinstance(fetcher_factory, Fetcher):
+        if extra_fetchers:
+            raise ConfigurationError(f"Cannot pass a Fetcher as 'fetcher_factory' with {extra_fetchers=}.")
+        if config:
+            raise ConfigurationError("Pre-initialized Fetcher given; section [fetching] redundant.")
+        return fetcher_factory
+
+    fetchers = []
+    if config:
+        fetchers.append(_make_fetcher(fetcher_factory, mapper_factory, **config))  # Add primary fetcher
+
+    fetchers.extend(
+        _make_fetcher(fetcher_factory, mapper_factory, **toml.load(file_fetcher_file)["fetching"])
+        for file_fetcher_file in extra_fetchers
+    )
+    if not fetchers:
+        raise ConfigurationError("Section [fetching] is required when no pre-initialized AbstractFetcher is given.")
+
+    return fetchers[0] if len(fetchers) == 1 else MultiFetcher(*fetchers, duplicate_source_discovered_action="warn")
 
 
 def _make_default_translations(**config: Any) -> Tuple[str, Optional[InheritedKeysDict]]:
@@ -138,11 +178,8 @@ def _make_default_translations(**config: Any) -> Tuple[str, Optional[InheritedKe
 
 def _make_mapper(parent_section: str, mapper_factory: MakeMapperType, config: Dict[str, Any]) -> Optional[Mapper]:
     if isinstance(mapper_factory, Mapper):
-        if "mapping" in config and mapper_factory != default_mapper_factory:
-            warnings.warn(
-                f"Section [{parent_section}.mapping] in the configuration is "
-                f"ignored since a pre-initialized Mapper was given."
-            )
+        if "mapping" in config:
+            raise ConfigurationError(f"Pre-initialized Mapper given; section [{parent_section}.mapping] redundant.")
         return mapper_factory
 
     if "mapping" not in config:
@@ -157,18 +194,13 @@ def _make_mapper(parent_section: str, mapper_factory: MakeMapperType, config: Di
     return mapper_factory(config, for_fetcher)
 
 
-def _make_fetcher(factory: MakeFetcherType, mapper_factory: MakeMapperType, **config: Any) -> Fetcher:
-    if isinstance(factory, Fetcher):
-        return factory
-    elif not config:
-        raise ConfigurationError("Section [fetching] is required when no pre-initialized Fetcher is given.")
-
+def _make_fetcher(factory: FetcherFactory, mapper_factory: MakeMapperType, **config: Any) -> Fetcher:
     mapper = _make_mapper("fetching", mapper_factory, config) if "mapping" in config else None
 
     if len(config) == 0:
         raise ConfigurationError("Fetcher implementation section missing.")
     if len(config) > 1:
-        raise ConfigurationError(f"Multiple fetcher implementations specified: {sorted(config)}.")
+        raise ConfigurationError(f"Multiple fetcher implementations specified in the same file: {sorted(config)}")
 
     clazz, kwargs = next(iter(config.items()))
     kwargs["mapper"] = mapper